+++ internal/core/ports/cart.go
@@ -9,7 +9,7 @@ import (
 
 type CartService interface {
 	GetCart(ctx context.Context, userId uuid.UUID) (*domain.Cart, error)
-	AddItemToCart(ctx context.Context, userId, productId uuid.UUID, quantity uint8) error
+	AddItemToCart(ctx context.Context, userId, productId uuid.UUID, quantity int16) error
 	RemoveItem(ctx context.Context, userId, productId uuid.UUID) error
 	Clear(ctx context.Context, userId uuid.UUID) error
 }
